Pull the batch out to a separate service.
