+++ islandora.services.yml
@@ -59,3 +59,9 @@ services:
     arguments: ['@jwt.authentication.jwt']
     tags:
       - { name: event_subscriber }
+  islandora.upload_children.batch_processor:
+    class: Drupal\islandora\Form\AddChildrenWizard\BatchProcessor
+    arguments:
+      - '@entity_type.manager'
+      - '@database'
+      - '@current_user'
